+++ packages/sdk-gen/src/generate.ts
@@ -1,161 +1,73 @@
-/* eslint-disable no-console */
-require("dotenv").config();
-import { generate } from "@graphql-codegen/cli";
-import fs from "fs";
-import path from "path";
+import { generate } from '@genql/cli'
+import fs from 'fs'
+import path from 'path'
 
-import { Args } from ".";
-import { getBGsdkConfig } from "./util/get-b-gsdk-config";
-import { getBGsdkDirectoryPath } from "./util/get-b-gsdk-directory-path";
-
-export const DEFAULT_DIR_NAME = "sdk-gen";
+import { Args } from '.'
+import { getBGsdkConfig } from './util/get-b-gsdk-config'
+import { getBGsdkDirectoryPath } from './util/get-b-gsdk-directory-path'
 
 export async function main(args: Args) {
-  const bgsdkDirectoryPath = getBGsdkDirectoryPath(
-    process.cwd(),
-    args["--dir"]
-  );
+  console.log('Generating...')
 
-  if (!bgsdkDirectoryPath) {
-    throw new Error(
-      `Make sure you have a "${DEFAULT_DIR_NAME}" directory in the root of your project.`
-    );
-  }
+  const bgsdkDirectoryPath = getBGsdkDirectoryPath(process.cwd(), args['--dir'])
+  const config = getBGsdkConfig(bgsdkDirectoryPath)
 
-  const config = getBGsdkConfig(bgsdkDirectoryPath);
-
-  const codegenOutput = await generate(
-    {
-      schema: {
-        [config.endpoint]: {
-          headers: config.headers,
-        },
-      },
-      generates: {
-        [__dirname + "/generated/index.ts"]: {
-          documents: [
-            bgsdkDirectoryPath + "/**/*.{gql,graphql}",
-            bgsdkDirectoryPath + "/*.{gql,graphql}",
-          ],
-          plugins: [
-            "typescript",
-            "typescript-operations",
-            "typescript-generic-sdk",
-          ],
-          config: {
-            dedupeFragments: true,
-            useTypeImports: true,
-            documentMode: "string",
-            ...config.parameters
-          },
-        },
-        [__dirname + "/generated/graphql.schema.json"]: {
-          plugins: ["introspection"],
-          config: {
-            documentMode: "documentNode",
-            withHooks: true,
-          },
-        },
-      },
+  await generate({
+    endpoint: config.endpoint,
+    headers: {
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+      ...config.headers
     },
-    false
-  );
+    output: path.join(bgsdkDirectoryPath, 'generated'),
+    verbose: args['--verbose'] ?? false
+  })
 
-  const schemaCodegen = codegenOutput[0].content.startsWith("{")
-    ? codegenOutput[0]
-    : codegenOutput[1];
-  const sdkCodegen = codegenOutput[0].content.startsWith("{")
-    ? codegenOutput[1]
-    : codegenOutput[0];
+  // extra generated
+  fs.appendFileSync(
+    path.join(bgsdkDirectoryPath, 'generated', 'index.ts'),
+    extraGenerated
+  )
 
-  createDirIfDoesNotExist(`${bgsdkDirectoryPath}/generated`);
-  fs.writeFileSync(
-    path.join(bgsdkDirectoryPath, "generated/graphql.schema.json"),
-    schemaCodegen.content
-  );
-  fs.writeFileSync(
-    path.join(bgsdkDirectoryPath, "generated/index.ts"),
-    "/* eslint-disable */\n" +
-      "// @ts-nocheck\n" +
-      sdkCodegen.content +
-      "\n" +
-      extraGenerated
-  );
-  const skdFilePath = path.join(bgsdkDirectoryPath, "sdk.ts");
+  // add eslint disable and ts nocheck to `types.ts` generated file
+  // at the beginning of the file
+  const typesFilePath = path.join(bgsdkDirectoryPath, 'generated', 'types.ts')
+  const typesFileContents = fs.readFileSync(typesFilePath, 'utf-8')
+  fs.writeFileSync(typesFilePath, `${topLevelAppend}\n${typesFileContents}`)
+
+  // sdk file
+  const skdFilePath = path.join(bgsdkDirectoryPath, 'sdk.ts')
   if (!fs.existsSync(skdFilePath)) {
-    fs.writeFileSync(skdFilePath, sdkFileContents);
+    fs.writeFileSync(skdFilePath, sdkFileContents)
   }
 
-  console.log("Done ✨");
+  console.log('Generated ✨')
 }
 
-function createDirIfDoesNotExist(p: string) {
-  if (!fs.existsSync(p)) {
-    fs.mkdirSync(p);
-  }
+const extraGenerated = `
+import type { Config } from '@bsmnt/sdk-gen'
+
+type CreateSDKParams = Omit<Config, 'headers'> & Omit<ClientOptions, 'url'>
+
+export const createSdk = function (params: CreateSDKParams) {
+  return createClient({
+    ...params,
+    headers: {
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+      ...params.headers
+    },
+    url: params.endpoint
+  })
 }
-
-const extraGenerated = `import type { Config } from "@bsmnt/sdk-gen";
-
-const fetch = globalThis.fetch || require("isomorphic-unfetch");
-
-type ClientOptions = {
-  noThrowOnErrors?: boolean;
-};
-
-export const createSdk = ({
-  endpoint,
-  headers,
-  clientOptions,
-}: Config & { clientOptions?: ClientOptions }) => {
-  const client: Requester = async (doc, vars, options?: ClientOptions) => {
-    const allClientOptions = { ...clientOptions, ...options };
-
-    const response = await fetch(endpoint, {
-      method: "POST",
-      headers: {
-        accept: "application/json",
-        "Content-Type": "application/json",
-        ...headers,
-      },
-      body: JSON.stringify({
-        query: doc,
-        variables: vars,
-      }),
-    });
-    const json = await response.json();
-    const { errors, data } = json;
-
-    const hasErrors = errors && Array.isArray(errors) && errors.length > 0;
-
-    if (hasErrors && !allClientOptions?.noThrowOnErrors) {
-      const message = \`GraphQL fetch errors:
-
-      \${errors.map((e: any, idx: number) => \`\${idx}. \${e.message}\`).join("\\n")}
-      
-      ——————
-
-      Doc:
-      \${doc}
-      
-      Vars:
-      \${vars}
-      \`;
-
-      throw new Error(message);
-    }
-
-    return { ...data, ...(hasErrors ? { errors } : {}) };
-  };
-
-  const generatedSdk = getSdk(client);
-
-  return { ...generatedSdk, client };
-};
-`;
+`
 
 const sdkFileContents = `import config from './config'
 import { createSdk } from './generated'
 
 export const bsmntSdk = createSdk(config)
-`;
+`
+
+const topLevelAppend = `/* eslint-disable */
+// @ts-nocheck
+`