+++ packages/sdk-gen/src/types.ts
@@ -1,1 +1,1 @@
-export type { BGsdkConfig as Config } from "./util/get-b-gsdk-config";
+export type { BGsdkConfig as Config } from './util/get-b-gsdk-config'